move command functions to separate files
